+++ main.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from rlmm.environment.openmmEnv import OpenMMEnv
 from rlmm.utils.config import Config
-from rlmm.rl.Expert import  ExpertPolicy,FastRocsPolicy, RandomPolicy
+from rlmm.rl.Expert import  ExpertPolicy, RandomPolicy
 import pickle
 import os
